git-find-efuncs.pl: support --private and --no-public
The default behavior has changed from "include private and public functions in
the output" to "only include public functions"; you can pass '--private' as an
argument to the script to get the old behavior back.

If '--private' is passed, then the '--no-public' argument can be used to
exclude public functions from the output. Note that '--no-public' is a no-op
unless '--private' is also passed.